change types in model and schema

+++ data-provider/src/main/scala/velocorner/model/Account.scala
@@ -31,7 +31,7 @@ object Account {
  * Represents a generic account used in the storage layer.
  */
 case class Account(
-  athleteId: Int,
+  athleteId: Long,
   displayName: String,
   displayLocation: String, // city, country
   avatarUrl: String,
